* stable-3.0:
  Broker publisher metric for number of message published
  Use KafkaConfiguration for Kafka related properties and functionality
  Fix setup.sh script and use bazel-bin directory
  Use bazel-bin to look for plugin's output
  Move ProjectDeletedSharedDbCleanup to global Module
  Decouple Kafka configuration class
  Add LB (HAProxy) to dockerised environment
  Add health-check plugin to dockerised environment
  Add multi-site plugin to dockerised environment
  Move CURATOR_VER constant to right before its use
  Fix plugins URL in the multi-site local environment
  Zookeeper module reads it's own configuration
  Update design proposal to abstract broker and Ref-DB
  Move Zookeeper-specifics into its own Module
  Don't import Id/Key/NameKey directly
  Module: Simplify BufferedReader creation
  DefaultSharedRefEnforcement: Fix MutableConstantField warning
  Fix ClassCanBeStatic warnings
  Fix DefaultCharset warnings
  ChangeCheckerImpl: Fix OperatorPrecedence warning
  Module: Fix DefaultCharset warning
  Configuration: Fix MutableConstantField warning
  KafkaSubscriber: Fix ClassCanBeStatic warning
  Bazel: Drop dependency on commons-lang3 library
  Bazel: Remove unneeded dependencies from test rule
  Bazel: Harmonize library names
  Format build files with buildifier
  Upgrade zookeeper to 3.4.14
  Update test dependencies
  Consolidate source of generated image files
  Allow restart of Gerrit instances
  Add header to the custom Kafka properties

+++ dockerised_local_env/gerrit-2/docker-entrypoint.sh
@@ -1,19 +1,19 @@
 #!/bin/bash -ex
 
-java -Xmx100g -jar /var/gerrit/bin/gerrit.war init -d /var/gerrit --batch --dev --no-auto-start --install-all-plugins
-#
-echo "Gerrit configuration:"
-cat /var/gerrit/etc/gerrit.config
-echo "Replication plugin configuration:"
-cat /var/gerrit/etc/replication.config
-
-echo "Remove git repos created during init phase"
-rm -fr /var/gerrit/git/*
-
 echo "Starting git daemon"
 /usr/local/bin/git-daemon.sh &
 
-echo "Waiting for initial replication"
-sleep 120
+if [[ $INIT == 1 ]]; then
+  java -jar /var/gerrit/bin/gerrit.war init -d /var/gerrit --batch --dev --no-auto-start --install-all-plugins
 
-java -Xmx100g -jar /var/gerrit/bin/gerrit.war daemon
+  echo "Remove git repos created during init phase"
+  rm -fr /var/gerrit/git/*
+
+  echo "Waiting for gerrit1 server to become available."
+  sleep 120
+  ssh -p 29418 admin@gerrit-1 replication start
+  echo "Waiting for replication to complete."
+  sleep 30
+fi
+
+java -jar /var/gerrit/bin/gerrit.war daemon
